Add files via upload
The HTMLParser class parses simple HTML files and create a DocumentObjectModel. The CSSParser class parses simple CSS files. Both take into account small subsets of the wide variety of syntax allowed in HTML/CSS and are quite error intolerant. We can see if we want to make these more complex or concentrate on other parts of browser (such as networking); currently I am through with the first three articles of the tutorial on rendering engine.